django/requirements.txt
Django==1.10.3
django-cors-headers==1.3.1
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
markdown2==2.3.1
mod-wsgi==4.5.7
mod-wsgi-httpd==2.4.23.1
mysqlclient==1.3.9
Pillow==3.4.1
Pygments==2.1.3
PyJWT==1.4.2
pyserial==3.2.1
pytz==2016.7
requests==2.12.4
